Implement "AssetDirs"
Per Krzyhau, it would be much better for us to let the user define multiple asset directories rather than just limiting them to one directory at a time.
The downside is that this makes it somewhat complicated to load assets, as now we have a bunch of assets pretty much everywhere. Woo hoo...